app/requirements.txt
tensorflow==2.3.0
numpy==1.18.0
keras==2.4.3
gensim==3.8.3
nltk==3.5
mysql-connector-python==8.0.23
pandas==1.2.2